made a nav component and linked it in the app.js
